Backend/requirements.txt
Django>=5.1.5
djangorestframework>=3.14
gunicorn
psycopg2-binary
dj-database-url
